resolved pull request comments
